clippy stuff

+++ crates/sui-graphql-rpc/src/types/event/mod.rs
@@ -143,7 +143,7 @@ impl Event {
 
                 let (prev, next, mut events): (bool, bool, Vec<StoredEvent>) =
                     if let Some(filter_query) =  query_constraint {
-                        let query = add_bounds(filter_query, &filter.transaction_digest, &page, tx_hi as u64);
+                        let query = add_bounds(filter_query, &filter.transaction_digest, &page, tx_hi);
 
                         let (prev, next, results) =
                             page.paginate_raw_query::<EvLookup>(conn, checkpoint_viewed_at, query)?;
@@ -177,7 +177,7 @@ impl Event {
                     } else {
                         // No filter is provided so we add bounds to the basic `SELECT * FROM
                         // events` query and call it a day.
-                        let query = add_bounds(query!("SELECT * FROM events"), &filter.transaction_digest, &page, tx_hi as u64);
+                        let query = add_bounds(query!("SELECT * FROM events"), &filter.transaction_digest, &page, tx_hi);
                         let (prev, next, events_iter) = page.paginate_raw_query::<StoredEvent>(conn, checkpoint_viewed_at, query)?;
                         let events = events_iter.collect::<Vec<StoredEvent>>();
                         (prev, next, events)
